Reanimate clang-tidy
Untangle CMakeLists.txt so tidy can run.
Disable clang-tidy by default (build with it takes ~40 minutes).
Redo the config - disabling checks that don't pass. They are to be fixed
in another task.
Exclude some directories that shouldn't be checked.
Fix a trivial bug and a hard one that was blocking tidy from working.

Change-Id: I18b27a90c0c67d5e0f756b1c7b477a3ad4973f61