ScatterAdd() : Set dtype of output metadata correctly.
Fix for: RuntimeError: Cannot report itemsize of Tensor that
doesn't have initialized dtype

This was caused because the dtype of output metadata was not set
for the final cast operator in the unsorted scatter add bf16 path.

Since cast is the last operator in the unsorted scatter add bf16 path,
we need to use the incoming output_metadata for the
cast_op3->AllocateAndAddSynapseNode. At the same time, cast operator
needs the dtpe to be set in output meta data. dtype may not be set in
the incoming output_metadata if ScatterAdd operator is used as an
intermediate op. Eg. in EmbeddingDenseBackward operator.
So, use the incoming output_metadata, but set the dtype correctly.

Also remove the environment variable used to WA the issue.

Change-Id: Ib65b07fb3cc5ddd89c7963647da27acd57c3fed1
Signed-off-by: Internal Developer <internal-developer@intel.com>