Fix for OOM issue using shared_future
Future object is stored and used only for the
collective and copy threads. To avoid race condition
of this future object in wait_for_future, used
shared future to use the copy of the object.
Using shared future in Execution thread causes
delay in the tensor release which is causing OOM
in some models.

Since the shared future is not mandatory for the
execution threads as we dont save the future object
and use it later. So removing the use of shared future
in execution thread.

Change-Id: Id10b2eb03cdd1ce677c58f2a331ad0fa30ca144b
Signed-off-by: Internal Developer <internal-developer@intel.com>