Increment the recipe counter, after submission
Recipe counter is incremented in the before
submission of recipes to synapse/hccl. When
memory is lower for the required job, then
it is freed by waiting for all the recipes
in queue to complete execution. In some case
the counter got incremented from collective
and graph, since the increment is before
submission, this caused incorrect wait condition.
So increment the recipe counter only when
the current recipe is submitted.

Change-Id: I3cf71053dad1edae0154766800e9a35f5b39f887
Signed-off-by: Internal Developer <internal-developer@intel.com>