Fix test failure with reset memory handles
synapseMemoryTest needs memory to be cleaned state,
so resetMemoryPool is done. Some tests that
are run previous to test there is a dealy in Tensor
release, these are delated later after the test completes.
If the tensor release come after reset handle which
done by synapseMemoryTest, there will failure as the
handles wont be found.
So ignore the tensor release when reset of handles
is done.

Change-Id: I4afc56c67c2e5081811ac0a308346a6c43c1c99f
Signed-off-by: Internal Developer <internal-developer@intel.com>