Enable generic stream by default
Current performance issue when enabling generic
is becasue for each stream synapse executes in
order, bascially completes one cmd and then only
it excutes the next cmd, even when there is no
dependencies. This has to be taken care in the bridge.

With repect to this default stream will create 4 synapse
stream.(D2H, H2D, D2D and Compute). if some cmd has to
be executed in parllel, user stream need to be created.

Change-Id: I419a3e54617621ea3fb2011e17f326626efd1168
Signed-off-by: Internal Developer <internal-developer@intel.com>