PT_HPU_ENABLE_SYNC_OUTPUT_HOST is false by default
Changing the default: PT_HPU_ENABLE_SYNC_OUTPUT_HOST=0
------------------------------------------------------------------------
This patch changes the default value of environment variable
PT_HPU_ENABLE_SYNC_OUTPUT_HOST from true to false, to prevent the bridge
from incurring an additional, redundant synchronization when the user
calls a collective operation with default parameters (assuming
async_op=True).

This is to match the behavior of Habana TensorFlow bridge and prevent
the user from paying an additional, completetly unnecessary cost of
synchronization after calling the collective (e.g. all_reduce).

Addressing an exit deadlock
------------------------------------------------------------------------
The patch comes with a fix to a bug occuring when processes exit just
after issuing a collective operation. In that case, the event completion
handler tries to deallocate the tensor (subject of a collective op), but
meanwhile the stack cleans up the HPU subsystem in ProcessGroupHCCL
destructor. Both processes are GIL-locked, which leads to a clash and an
infinite deadlock.

The issue has been addressed by delegating/deferring/postponing both
processes onto a separate thread using JobThreadHCCL, which itself waits
for all the jobs to be completed upon closure.

Thanks-To: bsochacki@habana.ai
Thanks-To: tkrawiec@habana.ai
Change-Id: Ia0c093c67720f8bd8682a25cf48e27fe8220e4b0