Correct seed tensor position for normal and log_normal
The above kernels have an optional tesor input before seed tensor
(also optional) input. Eg stddev tensor. If we are not passing this tensor
to TPC we should set it as null. This is because TPC requires that all
leading unused optional tensors are passed as null if any valid tensor
(eg. seed in this case) follows them.

Change-Id: I4c726d48eb16bff3e5340f3ae3d5111a2ffc3231
Signed-off-by: Internal Developer <internal-developer@intel.com>