Add support for handling consecutive D2H/H2D NB
When there is a D2H and the a immediate H2d copy(required
by some models deepspeed) and copy is a nonblocking and
not using pinned memory, Bridge needs to register producer
events for D2H so that H2D will add wait event.
In addition to this, in case of non-pinned memory, there
is a std:copy from pinned memory to host, so this
also needs to be handled to make sure copy is completed
from host side. So, add support in the bridge to handle
this case by registering device events and host event
to handle the std:copy.
Ex:
t1 = torch.arange(1, 5, dtype=torch.bfloat16, device='hpu:0')
t2 = torch.zeros(4)
t2.copy_(t1, non_blocking=True)
t2 = t2.to(t1.device)

Change-Id: Ic9fdab6e2f8f74b42b623b4a5ab33837acda33a0
Signed-off-by: Internal Developer <internal-developer@intel.com>