Add user Event to map to 4 synapse events
Default streams are mapped internally to 4 synapse
streams. so add a user event to these stream it
needs to mapp to 4 synapse events. to make the
interface generic, to any user event internally
map to 4 synapse stream. This would mean any
API on a user event will need to map to all
4 synapse stream.

Add the above support and also cleaned up
HPUGuardImpl event function to use the HPUEvent.
This also fix the wait event for default stream.

Change-Id: Ib9838c5607a3b5a4330cbad846e61df898f21903
Signed-off-by: Internal Developer <internal-developer@intel.com>