return stream handle as uintptr_t instead of intptr_t
In PYBIND 'compute_stream' function the return value is casted to intptr_t
and then reinterpreted again as uintptr_t on the Media stack.
This is causing BYBIND to fail the cast between python code and C++ with error:
RuntimeError: Unable to cast Python instance to C++ type.
This issue happened only if the 64'th bit of the stream handle is '1' which represent
the sign on this case.
commit https://gerrit.habana-labs.com/#/c/328276/ introduced a new infra to handle
streams which uses the last bits of the 64 bit number and exposed this bug.

Change-Id: Id087165bfa947efbfe4632090a7f8deaf652c7a9
Signed-off-by: Internal Developer <internal-developer@intel.com>