Fused SDPA:Move seed generation to FE
Seed generation was earlier done in python wrapper using
hpu.random.seed(). But this does not honour torch.manual_seed()
So the FSDPA output was different in two runs even when the
test script set the same manual seed for both runs.

So this patch moves seed generation to BE and uses
the C++ generator.random() to generate seed as done
in random data ops

Change-Id: I9dec744bec081d40519a40c57a934f96cc8108dc
Signed-off-by: Internal Developer <internal-developer@intel.com>