SliceOperator - Fix end parameter not specified
End parameter type validation was causing an error in cumprod backward.
In pytorch's implementation cumprod uses slice without end parameter,
then in PT Bridge it is marked as None. In this case, pytorch's slice op
sets the end parameter as INT64_MAX and then aligns it to the maximum
value of the self's shape on axis=dim. This was missing in our
implementation. Alignment is done in the if block
"needs_params_handling".

Change-Id: I340c705681fa1f7aa752a2fa6e33a8c746f480c3