Fix for defrag test failure
Running defragmenter test needs memory to be in
clean state. Some of the test which ran before this
has tensor that are not released. When memory is reset
and free of tensor occur later these handles are not
found in the map and causes assert.So, mark each handle
that is not freed to reset and ignore when the handle
state is reset.This will be called only in case of tests.

Change-Id: I29f747d5262e7d1013080240d3dd3fb223ba879f
Signed-off-by: Internal Developer <internal-developer@intel.com>