Change InstanceNorm OP impl to use instance_norm cguid in eager and compile mode
Before this commit, the InstanceNorm OP was decomposed into batch_norm kernels which are not deterministic. Consequently,
the results obtained in lazy and eager/compile differed. With this change, the InstanceNorm uses always dedicated
instance_norm kernel in each mode, producing deterministic results. As a part of this task, the instance_norm kernel was
also moved to backend. Furthermore, the _custom_op_meta_registrations script is imported when initializing torch.compile.

Change-Id: Ieaaeec0b98c23daf20d7385e64824b8ae6ff349c