Remove remainder.Scalar_Tensor frontend
There was an error in eager "C++ exception with description
'aten::remainder.ScalarTensor is not yet supported on HPU'"
It was caused by the fact that a custom frontend was created for lazy,
but not for eager. The lazy frontend was empty, so instead of adding
eager frontend, it would be enough to remove the empty lazy frontend.

Change-Id: Ia9c46e64ff283805e6692a7b868761d32b115972