Use meta StridedView for eager ops not using input tensor
For eager ops like fill_/zero_, the input tensor is not added to Synapse graph
and only the shape is used by these ops. When the input is a StridedView node,
this creates a StridedView node in Synapse graph which is not connected to the
actual op. This causes a data race as StridedView and the op works on same memory
without control edge.

The fix is to not add StridedView to Syanpse graph for such scenario, by marking
the StridedView node with a meta JIT attribute.

Co-authored-by: Internal Developer <internal-developer@intel.com>

Signed-off-by: Internal Developer <internal-developer@intel.com>
Change-Id: I0fa8d86ff797caafe2f7b8e8ca9f6409aa4cf4c2