Fix graph compile error for scalar tensor input
When the input is a scalar tensor, the dim() returns 0 for the tensor.
A few kernels that uses self.dim() - dim - 1 to caculate the tpc axis
param will end with -1 which is invalid for the underlayer graph thus
caused compile error when compiling.

The fix is to call get_dim_in_tpc_order which is a util function do the
safe guard of the calculations considering the case of 0D tensor.

Change-Id: I2c7d1a50e92067e1db5f3a1b2b76695d685c0d54