Enable avg_pool1d in torch.compile
PyTorch decomposes avg_pool1d into unsqueeze -> avg_pool2d -> squeeze.
This change only apply to the case when avg_pool1d has an input with
rank 2. It is unsqueezed to rank 3, but avg_pool2d didn't support such a
case. This required fixing the output shape calculation in such a way
that we assume that the dimension determining the number of batches has
size 1, and it also required expanding the input to 4d and the output
from avg_pool2d kernel to 3d.

Change-Id: Ib33fa9f3d3856a813852a3031aa13e8bca2d9718