Exception handling for pipeline threads
Currently the exception raise from compile and execution threads
are caught and ignored. The lowering thread and main thread
continues with wrong output. The test cases experiencing such issues
shown as result mismatch.

The fix is to log and rethrow the exeception to pop error up.
We also change to use simple future and reset the thread handle
when join is called. The join call needs to be guarded by a lock.

Change-Id: I5ec2e36487347ae21797105212601bce6b779975