Update kernel parameter settings for Tril/Triu
The existing Tril/Triu kernel parameter setting depended on
input shape. This caused issues with dynamic shape where, on a
cache hit, the parameters set for the kernel were for the
previously compiled shape.

For the Pytorch definition of Triu/Tril, it is enough to set the
numUpper/numLower parameters as max positive and -ve int values
respectively. Hence the shape dependency is avoided.

Change-Id: I4adc1dca473d8b0c1228962d1ce75afd8c5251c2
Signed-off-by: Internal Developer <internal-developer@intel.com>