Fix _fused_dropout in torch.compile
This commit adds output meta function for _fused_dropout op and fixes
a problem with the generator input. In torch.compile there is no frontend
that converts the generator to tensor, so in this case getNextInput for
the 3rd input throws an error regarding an unexpected input type.
To fix it, the 3rd input should be checked whether it is of type Tensor
and if not, then the seed should be taken by calling syn_seed().

Change-Id: I410db0eef53d3e43c3f8d5ff29026669d856465b