Fix div output mismatch when rounding is used
The problem is the floating point divide algorithm used by TPC div
f32 kernel is not so precise. For example the result of 3/3 is
0.9999999404 instead of 1.0. In this case, when rounding mode such
as floor or trunc is used, the result is 0 instead of 1. Similar cases
exist for other div scenarios.

A new div kernel called div_precise for f32 is implemented which is
target to be more precise than the existing div f32 kernel but of cource
with somewhat lower performance exchanging for accuracy.

The solution is to use the precise version for some scenarios.
Considering the performance and accuracy aspects of the two div kernel,
we use a env flag to allow user to enforce div precise. For normal
div operators, if user doesn't set the flag, the performance version
will be used by default. And considering that accuracy is especially
important for rounding cases, the precise version will be used by
default if user doesn't set the flag.

Change-Id: I1f85a2f808780fdc54d08109ba20d9e94d9241a7