Fix RecipeCacheLRU data corrupt issue
RecipeCacheLRU is possible by accessed in multiple threads context but
its clear method is not guarded by lock as its add and drop_lru methods.

The Jenkins task Test_pytorch_modules_gaudi_sim_cpp_suite_master_next
steadily report "lru cache corruption, map size 0 not equal to list_size 68"
exception on case nativeBatchNormLegitNoStatsTest which indicates that
the clear and add were called in different threads and caused data
corruption when clear is not in the lock.

The fix is to lock the clear() method as method.

Change-Id: I66fc64fb69a383ca870251cdd08cc6c9e196d6bb