Fix for collective ops usage with habana cache
When two identical graphs containing an allreduce were processed
one after another, a race between two allreduces occurred - one of
them wrote down its input and output addresses in the cache and
proceeded to process the reduce op while the next reduce began its
work and I rewrote its addresses into the cache. The entire structure
collective_kernel_info is written in the cache, which contains a
vector of input adresses input_tensor_infos, which led to the fact
that for two different operations the same structure was used with
the same vectors for storing inputs and outputs because of which
the addresses of the first ones were “lost” and operations were
performed on the tensor that could be written to memory last.
The problem solved by passing a copy of the input to the function
that processes it instead of passing its address.

Change-Id: I0ad21dbe7b14e16a0ff9f06dd8bf1c603f8c75a4