Fix aten.index for PT None
Implement advanced indexing for indices identified in Pytorch
as None. In C++ layer when calling aten.index PT API this value
is identified as an uninitialized optional variable or as an
initialized optional variable Tensor (undefined) when calling
indexing on self Tensor e.g. Tensor[[1, 0], None]

In current Pytorch implementation the test checking if the
indice is qualified for advanced indexing is a check if
the optional Tensor variables are uninitialized. The second part
checks if value in the initialized optional Tensor variable is
undefined. The first step of this flow was missing in Habana's
implementation. The check for advanced indexing was basing only
on the initialized optional variables and their values undefined.
This could only cover ther case with the self Tensor indexing
-> Tensor[[1,0], None]

The conclusion for the fix is that:
explicit Tensor indice is an initialized optional variable and
wrapped Tensor that has a defined value;
None and ":"(which is actually a slice(None)) are uninitialized
optionals or initialized otionals with Tensors that have no value

The fix was applied for eager and lazy. Compile must fallback
to eager since the output shape can't be deduced.

All the failing tests are restored and serveral new ones were
added.

Change-Id: If8201dd92616a83691952bbfc36980b8d9b61d66