all_to_all op should take correct adress as its output
AllocateSynapseInplaceOutput method takes the op's input address as the
output address, which should be done in cases of operations that perform inplace,
replacing the values ​​at the input address. However, alltoll is not one of these
and the address at which the result of the operation should be written is one
of the inputs located at position 1. That is why, instead of the correct answer,
the result of the operation was the same values ​​that were supplied at the input.

Change-Id: I17d14b2c48a6f136f6e7d325b9e103390d03dec5