Fix clamp output mismatch
When the input tensor is integer type and the min/max parameter is
a float, the output of clamp is incorrect. The root cause is when
filling the clamp parameters (FillClampParams), the min/max was
incorrectly converted to int by only checking the input tensor type
instead of the the promoted common type which is float.

Change-Id: I5918c85a84d5eb747afda8bf7837381d1800f5b0