Support AllGather for uneven output splits
A all_gather requires each process to gather other process' inputs into an
output tensor according to the ouput list provided. Internally,
dist.allgather will merge a list of hccl._broadcast_oop to implement a vector
all-gather in the case when the any shape is different in the output list.
Otherwise, it will perform a hcclAllGather as usual.

This change adds an out-of-place `_broadcast_oop` function to
ProcessGroupHcclBase. It allows broadcasting an input tensor and placing
the output in a separate output tensor.

Change-Id: I75babc2c29515b137639635f91229173e54948d6
Signed-off-by: Internal Developer <internal-developer@intel.com>
Signed-off-by: Internal Developer <internal-developer@intel.com>