Fix foreach_round output mismatch
The round function implements round half to even. So we need to set the
ROUND_HALF_NEAREST_EVEN flag in the params. The original Foreach backend
implementation doesn't consider custom fill params and we add FillParams
call when building the node. It doesn't cause problem if there is no
params.

Change-Id: Ic400aff3c819f5e35f8d9fd4bddaff2a2ae0b7d3