Use torch.use_deterministic_algorithms to enable HPU determinism
Previously custom flag htcore.hpu.setDeterministic was available to do
this. Allow to set the behavior with torch standard way instead of
custom.
The old custom way is to be removed in 1.15 

Also, add test for batch norm to verify if determinism is correctly
enabled (skipped as work only locally run separately due to flags).

Change-Id: Ib1f8320bcd158714e9e4475f1116e52422cfabf1