Fix the cast error of foreach binary
Foreach binary used at::ScalarType for checking whether cast is needed
between two data types. This checking is incorrect for some data types
sharing the same underlayer type, such as double and float.

The fix is to use the cast type to check. ForeachAdd and ForeachMul were
failed and disabled for this reason. And now can be restored with this
fix. No addtional test cases are needed.

Change-Id: I0082ade65249fb31eef176a8f25070fe670192e3