[Eager] Fix for segmentation fault in FSDP with pipeline enabled
When "cat" op is followed by a "resize to zero" in
pipelining, JoinPendingPipelineThreads() is needed
to eliminate the race condition on the size of the
same tensor being used by both the ops. Also in
"set" op, we have to set is_tensor_pipelined_ flag
in set_source_Storage_storage_offset(), as this
op does not go via EagerOpBase::run().
Test case: Added test_eager_fsdp.py and tested it
locally on 8x setup as well.

Change-Id: I47e3d5b9d5957f47e739c28c8d9598c13cd5966a
Signed-off-by: Internal Developer <internal-developer@intel.com>