Reverting the shape change of 0-dim input tensors from 1-dim
During the first graph compilation, InferOutputMeta for IdentityOperator
changes the shape of zero-dimensional tensor to one-dimensional. This
causes an shape assertion failure in the dynamic shapes flow when in some
future iteration if there is a JIT cache hit, the InferOutputMeta method
is not invoked and the input tensor remains zero-dimensional.

Change-Id: Ic110cdea518ceef617a08b4ca7786b5cf6c7d468