Implement _unique eager
The implementation is based on nonzero_eager.

_unique is a non-inferable op. This means, that the output shape is not
known until the kernel is executed. This requires for some operations
(such as slicing and flipping) to be moved to frontend, as there is a
kernel execution present there. This kernel execution in frontend
produces an output tensor which is "back-padded" with zeros. In order
to aqcuire the tensor of the correct shape - which can be passed further
to backend - the zeros need to be sliced from the output tensor.

Due to torch::_unique producing nondeterministic results with the sorted
flag set to false, the tests handle this case differently. This has been
explained within the tests file itself

Change-Id: Ib25211d5f10f8f246e588fe971f8f45bd73e68c6