Fix optimizer_sgd_bwd call
The optimizer_sgd_bwd kernel supports the first input with bf16 and f32
type but there is a guid only in the f32 variant. The current call took
the type from the first input, which caused a non-existent guid to be
called. To fix this, the kernel call type was set to f32.

Change-Id: I8a3b8668e45c8b7f27edbd56d507951724208856