Add support for f16 dtype in randperm OP
This commit adds support for f16 dtype in randperm OP in all modes.
Furthermore, it fixes randperm lazy OP implementation, so the
random_shuffle CGUID is called only in int32 precision and not in OP
precision. If the dtype of the inputs is different, they are autocasted
on CGUID level to int32. Then the output is casted to the OP's dtype.

Change-Id: Iad729ac7524380d7736ddafac7b15f0a48c173cb