Preventing stat dump overwrite
If for a graph with some recipe_id,
dynamicity is not detected, and a fresh
compilation occurs, then any pre-existing
stat json file nwith the same recipe_id in its
name used to get overwritten with the
stats of this new compilation.

After this patch, in such scenarios,
a new file will be created with the
same recipe_id at start of its name,
followed by the device rank and the
graph_hash_code.

Example for new file name format:
In graph_recipe_1002000_1002000_5_12345.json,

1) "graph_recipe_1002000_1002000" is the
SynapseGraphName/recipe_id.

2) "_12345" i.e the number after the last
underscore is the gaph_hash_code.

3) "_5" i.e the  number after the second-last
underscore is the device node rank. For a single
node device, it defaults to 0.

Change-Id: I8e0960c012a6a5c9288f904f2d3cb670f8d6d7f7