Gaudi3 BGEMM takes longer at PT exec
BGEMM takes longer at PT exec relative to
synrec on Gaudi3. This is due to the incorrect
device memory alignment for Gaudi3 which is
expected to 256 instead of 128. Instead of
fixing the alignment to some value, query
the alignment info based on device type from
sync with synDeviceGetAttribute API.
Fix is to pass the alignment information to
device memory manger and use it for memory
allocator.

Already test exist for, this should pass on
gaudi3 also.
tests/cpp/synapse_helpers_memory_test.cpp

Change-Id: I3e2d6a6d8264e66657098bc550f9763bc54f3036
Signed-off-by: Internal Developer <internal-developer@intel.com>