[PyTorch/HuggingFace/BLIPModel] torch.full() returns an unexpected datatype (torch.float32)
The issue reproduced only in 'lazy' mode (does not occur in eager or
compile).
The CPU torch::full() was observed to use the 'fill_value' scalar
input type as the output type, when the 'dtype' optional parameter is s
et to 'None'. The HPU implementation was modified to do the same.

The 'dtype' input parameter is optional for torch.full(). If the 'dtype'
optional is set to 'None', a global data type should be used. This global
type is set to be a float32 or float64 (other types are not guaranteed to
function properly) by the 'torch.set_default_dtype()' function . The
previous implementation was resolving the global type properly (to
float32), which was not consistent with the actual behaviour of the CPU
implementation.

Change-Id: If359ae7efa3eb1ebf071fa02e336182b08a7a9f2