Fix GIL deadlock caused by blocking data copy operations
Random DFA was observered from a random Rank during Llama 13B
training with 8 cards. We tracked down the lock chain and found
out the deadlock caused by GIL.

One of the collective Done callback needs to clean up the related
resources among which are input and output tensors. The destructor
of these torch tensors will need acquire GIL. While the GIL is held
by the main thread which is blocked on the while loop for waiting
for the copy operation to complete. To complete the copy operation,
the Done callback needs to be called. The calling of the Done callbacks
are guarded by a shared lock in stream_event_manager. Since the collective
operation Done callback holds the shared lock and waiting for acquiring
GIL. This caused the dead lock.

The solution is to release the GIL before entering while loop waiting.
And acquire it after exit the while loop.

Change-Id: I017cfd7f48b92866d065331efb3fecdcf0ecc85e