Improve OP info logging for dynamic shape analysis
The original OP info logging in BuildSynapseGraph doesn't distinguish
the real compilation and shape inferences (min, max, output, hybrid).
This makes the OP info shown in the logging are not so accurate to
reflect the real uses.

This improve the OP info logging with the information of what it is
doing when calling into BuildSynapseGraph (or HybridSif). Specifically,
We prefix the logging with pass information when it is doing shape
inference: SIF_MIN for min pass, SIF_MAX for max pass, SIF_OUTPUT for
output or hybride sif pass.

This additional information can be utilized in the OP analysis tool
for summarizing a specific OP types for MIN, MAX and OUTPUT SIF.

We also logging the OP placement information and its dynamicity so that
we can get a list of original OPs that puts to eager or HPU cluster,
or belongs to static or dynamic for analyzing.

Also we make sure the OP validator for shared layer output the proper
operator name (aten operator name) when an OP is not supported. And
make sure the OP fallback check output proper formatted messages. This
information can also be utilized for fallback analysis.

Change-Id: If9c05dfdeaec574f56aff6610d34379598204070