Fix mod_exp function for INT64_MIN input value
Implementation of mod_exp function uses an input of dtype int64_t for
which abs value is calculated using llabs. This function calculates
absolute value of any long long input. Yet, it has a limitation that for
INT64_MIN (value for which abs cannot be represented as a long long int)
undefined behavior is caused. Thus when an input to mod_exp function
was INT64_MIN, the program didn't behave correctly. As a solution, the
abs value of INT64_MIN is calculated manually and stored in a uint64_t
typed variable.

Change-Id: I4e8b09a7e04d0cef8e536934700492f9f5cf99df