Override _to_copy decomposition for handling FakeTensor wrapping
PR 104689 handled the FakeTensor wrapping and merged since PT 2.1.
But it caused some new issues (such as 118790), the part of its code
to handle the FakeTensor wrapping for recording the real device of
the tensor got reverted by PR 119868 in pytorch 2.3.0 to fix these
issues. This caused the original issue Habana team tried to resolve
reopen. Further more, Habana pytorch integration has some code in
compile mode for handling the behavior of PR 104689 and is not working
on PT 2.3.0 and caused some new issues as a result of the revert.

This patch override _to_copy in torch._decomp.compositions with our
own version with the FakeTensor wrapping logic. _to_copy in
torch._decomp.compositions is a very core and internal part of torch.
what we need to override here is meta dispatch key for _to_copy.

The patch includes 2 key places to make the override work:
1. In compositions, we implement our version of _to_copy and
registering to our own  decomposition table.
2. In meta registrations, we update the python kernel registrations
for Meta dispatch key of aten::_to_copy operator with our version.

torch._subclasses.fake_tensor.torch_decomp_decompositions overwrite
is no longer needed because we don't override directly the global
composition table now.

Change-Id: I92daa4b0c022be4b5d02dce776ba513ffe4b8605