Disable fold quantize for convert_pt2e
In 2.3.0, pytorch quantize_pt2e turn on fold_quantize (constant fold)
by default. In 2.2, it is by default off.

When turn on fold_quantize, the constant tensor (such as module
parameters that static) will be optimized to be statically quantized.
And there is only dequantize node in the graph when needed, thus the
number of quantize nodes is not equal the number of dequantize nodes.

These broken all the logic in the implementation of decompose_quant_ops
which assumes the same number of quantize and dequantize nodes for
mapping the src and dst dtype needed by decomposition. And the test
cases also assumes and asserts the resulted operator numbers which are
based on the number of quantize and dequantize nodes.

It is not easy to handle the data type mapping enabling fold quantize.
To solve this issue, the solution is to explicitly disable fold_quantize
so that the behavior is the same as PT 2.2.

Change-Id: I9995db18b51ee3308ab677c22f27c78e3f71e93c