Batch Norm Fwd inf output meta fix
The change rearranges position for the last inferred meta output tensor
for the BatchNormFwd Op in the Output Inf mode.

The target shape for the first output from the BatchNormFwd Op is calculated after the op is "executed".
It happens because the kernel accepts 4D tensors, so reshape of the input tensor takes place,
4D input tensor is passed to the kernel (here, in the BuildNode func update for the Inf Output Tensors list happens,
but not for the first output, it is because final_result_index is not given, as we want other shape after all),
after that reshape to the original input shape is performed (here, the BuildNode func adds proper first output shape
as last element to the Inf Output Tensors list).
Consequently, we have to move last element of the Inf Output Tensors list at the beginning.

Change-Id: I4492a78aaf2e859b657785b11136d967e2325365