Change scatter.value backend to use ScatterValue CGUID
In the current implementation of the scatter.value OP, the constant tensor
is created, filled with the value passed in the scalar input. Such flow,
disables the shape agnostic flow as shape tensor to constant node is
created. To fix that, the dedicated CGUID was created, ScatterValue,
which handles the createion of the constant node on the complex_guid
level.

Change-Id: I412a7de530ac3578b30463acb682468f12d17a40