Fix kernel override warning for aten::instance_norm
The fix introduces an `override_instance_norm` function and an `override_composite_ops`
context manager to manage kernel overrides in PyTorch. The `override_instance_norm`
function checks if the input tensor is on the "hpu" device and applies a custom
implementation if true. Otherwise, it temporarily removes the current override,
calls the original implementation, and then restores the override. The
`override_composite_ops` context manager handles saving the original implementations,
applying the overrides, and restoring the original implementations once the context
is exited.

Change-Id: I26094dc2b9e1d5483061ff1081880329ed651368