Fix output mask for Convolution Backward op
* Hardcoded output mask to true for
  convolution_backward_overrideable op
* Skipped a test checking output_mask

For Hubert model, runtime error (no storage tensor attached to
lazy tensor) is observed without this patch.

This is a workaround for a bridge limitation with jit disabled
ops as it is a global list resulting in SIF fallback for all ops
of a particular type if one of the ops with the type is added to
the list due to failed InferOutputMeta.

Change-Id: Id6fb8f140094ce1bbb399ddc7a59ec031b781279
Signed-off-by: Internal Developer <internal-developer@intel.com>