Index PT integration update
Implement CGUID flow for eager index op.

Compile fallbacks to eager.

Lazy stays the same as dynamic flows do not
split the graph the same way it happens in backend,
synapse requires SIF for CGUID and there is none.

CGUID does know the size of the indices array
from the parameter, fix missing update in the
bool indexing.

Change-Id: I8ab85c0d8548bea672ea3b167ef8c77b0044f21f