Fix dtype of some Inplace params of optimizer_lamb_phase1 (fix for bf16 tests)
The dtype of input params wt_norm, adam_norm, and adam_step were not
inferred from params type when created. This led to the output type of
some syn_tensors to be misinterpreted at kernel building.

This fix uses the proper params type for the tensors mentioned and that
fixes the bf16 SingleOp tests

Change-Id: Idef5b64f211dfb7dcd8b09d090fb25ce72658551