Fix graph capture issue
In autograd evaluate() checks for the node
stream and sets it to that stream in
backward. since the graph capture is done on a
different stream, in this case set_stream is
called and this triggers mark_step, which
in turn calls mark Step in hpu and here the
check fails. Fix is to set the stream when
capture is in progress

Change-Id: I594be6f6bb3c33acf1cae9d2fd9d885081cf0755
Signed-off-by: Internal Developer <internal-developer@intel.com>